Add files via upload
The concept of the Efficient Frontier is essential in the world of finance and arguably the foundation of quantitative finance. But what exactly is this magical concept?

It is a scatter plot that illustrates the trade-off between the rate of return and the volatility (or risk) of a portfolio. It answers the question: How much risk (i.e., fluctuation in closing prices) are you willing to bear for the returns a portfolio can bring over a period of time?

In finance, we simulate thousands of portfolio combinations with different weights to visualise this trade-off. Magical, isn't it?

So, when you see a dot on the top right of the scatter plot, it likely represents a portfolio with higher risk (more variability in returns) but also higher expected returns. This reflects another foundational concept: "The greater the risk, the greater the potential reward."

Hope this makes sense!